requirements.txt
Django==1.10.3
psycopg2==2.6.2
djangorestframework==3.5.3
requests
dicttoxml==1.7.4
celery
redis==2.10.5